proof it works

+++ Native/src/main.cpp
@@ -109,7 +109,7 @@ void writeBlock() {
     uint8_t data[16];
     Serial.readBytes(data, 16);
 
-    auto status = mfrc522.MIFARE_Write(block[0], data, 16);
+    auto status = mfrc522.MIFARE_Write(block[0], data, (byte) 16);
     Serial.write((byte) status);
 }
